TAP interface: Delete existing route rules instead of just adding new ones
The different metric values can render the gateway not reachable.
It is more robust to delete the original rule and write a new one.

Note: dhclient and wpa_supplicant (see ss -0pa) still use the original interface
which is blocked by macvtap. When DHCP or authentification timeouts are hit after
some hours, the connection may break.